mapping/requirements.txt
earthengine-api>=0.1.400
geemap>=0.32.0
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
